requirements.txt
requests
ipykernel
Pillow
moviepy
matplotlib